+++ swiyu-ecosystem-portal-service/src/test/java/ch/admin/bj/swiyu/app/infrastructure/web/controller/AppConfigControllerIT.java
@@ -3,14 +3,14 @@ package ch.admin.bj.swiyu.app.infrastructure.web.controller;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import ch.admin.bj.swiyu.app.api.AppConfigDto;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import tools.jackson.databind.ObjectMapper;
 
 @ActiveProfiles("test")
 @AutoConfigureMockMvc