+++ swiyu-ecosystem-portal-service/src/test/java/ch/admin/bj/swiyu/app/infrastructure/web/controller/RegistrationControllerIT.java
@@ -6,20 +6,20 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import ch.admin.bit.jeap.security.test.WithJeapAuthenticationToken;
 import ch.admin.bj.swiyu.client.business.internal.api.BusinessPartnerApi;
 import ch.admin.bj.swiyu.client.business.internal.model.BusinessEntity;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.UUID;
 import org.junit.jupiter.api.Test;
 import org.mockito.Answers;
 import org.mockito.ArgumentMatchers;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import tools.jackson.databind.ObjectMapper;
 
 @ActiveProfiles("test")
 @AutoConfigureMockMvc